refactor: restore lib/solverCards.ts, address Arkana review
Reinstates the dedicated store module. `storage.ts` stays a pure
serializer façade; the subscriber state lives next to the cards.

From the Arkana review on #958: spell out that `switched` doubles as
`useCache` (a network switch may serve from the TTL cache, a card write
must bypass it), and say why `backupSolverCards` is deliberately absent
from the Solvers cleanup deps.
